src/backend/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Werkzeug==2.2.3
mysql-connector-python==8.1.0
scikit-learn==1.2.2
pandas==2.2.3
joblib==1.3.2
cryptography>=40.0.0
reportlab>=3.6.0
Flask-Session>=0.5.0
